ref: add command to remove benchmark files

+++ benchmark.sh
@@ -1,5 +1,17 @@
 #!/bin/bash
 
+cleanup() {
+    echo "Cleanup...."
+    [ -e "$INSERT_FILE" ] && unlink "$INSERT_FILE"
+    [ -e "$FIND_FILE" ] && unlink "$FIND_FILE"
+    [ -e "post.lua" ] && unlink "post.lua"
+    [ -e "get.lua" ] && unlink "get.lua"
+    echo "Cleanup completed"
+}
+
+# trap for cleanup on signals
+trap cleanup EXIT SIGINT SIGTERM
+
 # Generate random numbers for insert and find operations
 INSERT_FILE="inserts.json"
 FIND_FILE="finds.txt"
@@ -24,7 +36,10 @@ echo "Inserting numbers into the server..."
 while IFS= read -r line; do
   index=$(echo $line | sed 's/.*"index": \([0-9]*\).*/\1/')
   value=$(echo $line | sed 's/.*"value": \([0-9]*\).*/\1/')
-  curl -s -X POST -H "Content-Type: application/json" -d "$line" http://localhost:8080/v2/numbers/$index/$value > /dev/null
+  if ! curl -s -X POST -H "Content-Type: application/json" -d "$line" http://localhost:8080/v2/numbers/$index/$value > /dev/null; then
+    echo "Error inserting $line"
+    exit 1
+  fi
 done < <(tail -n +2 $INSERT_FILE | head -n -1 | sed 's/,$//')
 
 # Create a script for wrk to use for POST requests
@@ -41,7 +56,10 @@ EOF
 
 # Benchmark the insert operation
 echo "Benchmarking insert operation..."
-wrk -t12 -c100 -d30s -s post.lua http://localhost:8080
+if ! wrk -t12 -c100 -d30s -s post.lua http://localhost:8080; then
+  echo "Error fibding on line $line"
+  exit 1
+fi
 
 # Create a script for wrk to use for GET requests
 cat <<EOF > get.lua
@@ -61,7 +79,8 @@ EOF
 
 # Benchmark the find operation
 echo "Benchmarking find operation..."
-wrk -t12 -c100 -d30s -s get.lua http://localhost:8080
+if ! wrk -t12 -c100 -d30s -s get.lua http://localhost:8080; then
+  exit 1
+fi
 
-# Cleanup
-rm $INSERT_FILE $FIND_FILE post.lua get.lua
+echo "completed successfully"
